+++ src/coordinator/mod.rs
@@ -1,0 +1,2 @@
+pub mod coordinator;
+pub mod node_accepted;
